+++ src/core/io/asset_io.c
@@ -56,7 +56,7 @@ void asset_io_convert_mesh(const char* name)
   int len = 0;
   SPRINTF(ASSET_PATH_MAX + 7 + ASSET_IO_NAME_MAX, path, "%smeshes/%s", core_data->asset_path, name_src);
   buf = (void*)file_io_read_len(path, &len);
-  buf_len = len;
+  buf_len = (size_t)len;
 // #else
 //   zip_entry_open(zip_meshes, name);
 //   {
@@ -75,12 +75,12 @@ void asset_io_convert_mesh(const char* name)
   // -- save into file_dest as binary --
 
   u8* buffer = NULL;
-  asset_io_serialize_mesh(&buffer, verts, arrlen(verts), indices, arrlen(indices));
+  asset_io_serialize_mesh(&buffer, verts, (u32)arrlen(verts), indices, (u32)arrlen(indices));
   ARRFREE(verts);
   ARRFREE(indices);
    
   SPRINTF(ASSET_PATH_MAX + 7 + ASSET_IO_NAME_MAX, path, "%smeshes/%s", core_data->asset_path, name_dest);
-  file_io_write_bytes(path, buffer, arrlen(buffer));
+  file_io_write_bytes(path, buffer, (int)arrlen(buffer));
   ARRFREE(buffer);
 }
 
@@ -122,7 +122,7 @@ mesh_t asset_io_load_mesh(const char* name)
   asset_io_deserialize_mesh(buffer, &verts, &indices);
   
   mesh_t mesh;
-  mesh_make_indexed(verts, arrlen(verts), indices, arrlen(indices), &mesh);
+  mesh_make_indexed(verts, (int)arrlen(verts), indices, (int)arrlen(indices), &mesh);
   ARRFREE(verts);
   ARRFREE(indices);
   
@@ -180,13 +180,13 @@ void asset_io_convert_texture_dbg(const char* name, const char* _file, const int
   // }
 
   buf = (void*)file_io_read_len(path, &len);
-  buf_len = len;
+  buf_len = (size_t)len;
   ERR_CHECK(buf != NULL && buf_len > 0, "texture '%s' requested in %s(), doesn't exist in the asset folder.\n\t path: %s\n -> [FILE] '%s', [LINE] %d", name, __func__, path, _file, _line);
 
   // -- load pixels --
   stbi_set_flip_vertically_on_load(true);
   int w, h, channels;
-  u8* pixels = stbi_load_from_memory(buf, buf_len, &w, &h, &channels, 0);
+  u8* pixels = stbi_load_from_memory(buf, (int)buf_len, &w, &h, &channels, 0);
     
   // for (u32 i = 0; i < w * h * channels; ++i)
   // {
@@ -194,12 +194,12 @@ void asset_io_convert_texture_dbg(const char* name, const char* _file, const int
   // }
   // -- convert --
   u32 buffer_len = 0;
-  u8* buffer = asset_io_serialize_texture(pixels, w, h, channels, &buffer_len);
+  u8* buffer = asset_io_serialize_texture(pixels, (u32)w, (u32)h, (u32)channels, &buffer_len);
   
   // -- write to file --
   for (u32 i = 0; i < strlen(name) - 4; ++i) { name_dest[i] = name[i]; name_dest[i +1] = '\0'; }
   SPRINTF(ASSET_PATH_MAX + ASSET_IO_NAME_MAX + 12, path, "%stextures/%s.tex", core_data->asset_path, name_dest);
-  file_io_write_bytes(path, buffer, buffer_len);
+  file_io_write_bytes(path, buffer, (int)buffer_len);
   
   // -- cleanup --
   stbi_image_free(pixels);
@@ -211,30 +211,30 @@ u8* asset_io_serialize_texture(u8* pixels, u32 w, u32 h, u32 channels, u32* buff
   TRACE();
 
   const int header_size = (sizeof(u32) * 3);  // w, h, channels
-  const int pixels_len  = w * h * channels * sizeof(u8);
-  *buffer_len = pixels_len + header_size;
+  const int pixels_len  = (int)(w * h * channels) * (int)sizeof(u8);
+  *buffer_len = (u32)pixels_len + (u32)header_size;
   u8* buffer;
   MALLOC(buffer, *buffer_len);
   
   // pack header into u8
   // w
-  buffer[0] = w >> 24;
-  buffer[1] = w >> 16;
-  buffer[2] = w >>  8;
-  buffer[3] = w; 
+  buffer[0] = (u8)(w >> 24);
+  buffer[1] = (u8)(w >> 16);
+  buffer[2] = (u8)(w >>  8);
+  buffer[3] = (u8)(w); 
   // h
-  buffer[4] = h >> 24;
-  buffer[5] = h >> 16;
-  buffer[6] = h >> 8;
-  buffer[7] = h;
+  buffer[4] = (u8)(h >> 24);
+  buffer[5] = (u8)(h >> 16);
+  buffer[6] = (u8)(h >>  8);
+  buffer[7] = (u8)(h);
   // channels
-  buffer[8]  = channels >> 24;
-  buffer[9]  = channels >> 16;
-  buffer[10] = channels >> 8;
-  buffer[11] = channels;
+  buffer[8]  = (u8)(channels >> 24);
+  buffer[9]  = (u8)(channels >> 16);
+  buffer[10] = (u8)(channels >>  8);
+  buffer[11] = (u8)(channels);
 
   // copy pixels into buffer after header
-  memcpy(buffer + header_size, pixels, pixels_len);
+  memcpy(buffer + header_size, pixels, (size_t)pixels_len);
 
   return buffer;
 }
@@ -280,12 +280,12 @@ texture_t asset_io_load_texture_full_path(const char* path, bool srgb)
   texture_t t;
   t.loaded = true;
   t.handle = handle;
-  t.width = w;
-  t.height = h;
-  t.channel_nr = channels;
+  t.width  = (int)w;
+  t.height = (int)h;
+  t.channel_nr = (int)channels;
   
   #ifdef EDITOR
-  int t_path_len = strlen(path);
+  int t_path_len = (int)strlen(path);
   char* tex_name = (char*)&path[t_path_len - 1];
   for (int i = t_path_len - 1; i >= 0; --i)
   {
@@ -323,12 +323,12 @@ texture_t asset_io_load_texture_full_path_formatted(const char* path, bool srgb,
   texture_t t;
   t.loaded = true;
   t.handle = handle;
-  t.width = w;
-  t.height = h;
-  t.channel_nr = target_channels;
+  t.width  = (int)w;
+  t.height = (int)h;
+  t.channel_nr = (int)target_channels;
     
   #ifdef EDITOR
-  int t_path_len = strlen(path);
+  int t_path_len = (int)strlen(path);
   char* tex_name = (char*)&path[t_path_len -1];
   for (int i = t_path_len -1; i >= 0; --i)
   {
@@ -399,8 +399,8 @@ void asset_io_deserialize_texture_formatted(u8* buffer, u32 target_channels, u8*
   u32 d_pos = 0;
   u32 p_pos = 0;
   const u32 data_step      = MIN(target_channels, (*channels));
-  const u32 data_step_diff = MAX(0, (int)(*channels) - (int)target_channels);
-  const u32 pixels_step    = MAX(0, (int)target_channels - (int)(*channels));
+  const u32 data_step_diff = (u32)MAX(0, (int)(*channels) - (int)target_channels);
+  const u32 pixels_step    = (u32)MAX(0, (int)target_channels - (int)(*channels));
   // P_U32(data_step);
   // P_U32(data_step_diff);
   // P_INT(pixels_step);
@@ -448,7 +448,7 @@ void asset_io_serialize_archive(const char* dir_path, int initial_dir_path_len, 
 			strcat(path, "\\");
 			strcat(path, dp->d_name);		
       
-      int d_name_len = strlen(dp->d_name);
+      int d_name_len = (int)strlen(dp->d_name);
       // char* _name = path + initial_dir_path_len +1; // initial_... to cut off /assets/, +1 to cut off / 
 
       // @UNSURE: use as way to pack albedo, metallic, normal, roughness into one file
@@ -462,9 +462,9 @@ void asset_io_serialize_archive(const char* dir_path, int initial_dir_path_len, 
         PF("[.tex] \"%s\"\n -> \"%s\"\n", dp->d_name, path);
         int length = 0;
         u8* buffer = file_io_read_bytes(path, &length);
-        REALLOC(*rtn, (*rtn_size + length) * sizeof(u8));
-        memcpy(*rtn + *rtn_size, buffer, length);
-        *rtn_size += length;
+        REALLOC(*rtn, (size_t)(*rtn_size + (u32)length) * sizeof(u8));
+        memcpy(*rtn + *rtn_size, buffer, (size_t)length);
+        *rtn_size += (u32)length;
       }
       else if (type == ASSET_TYPE_MESH           &&
                dp->d_name[d_name_len - 5] == '.' &&
@@ -490,7 +490,7 @@ u8* asset_io_texture_write_pixels_to_buffer_u8(texture_t* t,  u32 gl_type, u32* 
 {
   TRACE();
 
-  *buffer_len = t->width * t->height * t->channel_nr * sizeof(u8);
+  *buffer_len = (u32)(t->width * t->height * t->channel_nr) * (u32)sizeof(u8);
   u8* buffer;
   MALLOC(buffer, *buffer_len);
   _glActiveTexture(GL_TEXTURE0);
@@ -519,9 +519,9 @@ void asset_io_texture_write_pixels_to_file(texture_t* t,  u32 gl_type, const cha
   u8* pixels = asset_io_texture_write_pixels_to_buffer_u8(t, gl_type, &pixels_len);
  
   u32 buffer_len = 0;
-  u8* buffer = asset_io_serialize_texture(pixels, t->width, t->height, t->channel_nr, &buffer_len);
+  u8* buffer = asset_io_serialize_texture(pixels, (u32)t->width, (u32)t->height, (u32)t->channel_nr, &buffer_len);
   
-  file_io_write_bytes(path, buffer, buffer_len);
+  file_io_write_bytes(path, buffer, (int)buffer_len);
 
   FREE(pixels);
   FREE(buffer);
